Improvements to runner and builder.
There were issues with the previous builder setup.

1. rust-analyzer was not working within the kernel directory
2. The artifact dependency was a bit wonky to say the least.

These issues are now solved. The builder moved to its own directory and bindeps
helps with managing the build.
