+++ builder/.cargo/config.toml
@@ -1,0 +1,2 @@
+[unstable]
+bindeps = true
